strands/requirements.txt
strands-agents
boto3
psycopg2-binary